refactor(entity): centralize component construction across spawn paths
Share one component installer across authored map, live and held construction and one runtime identity constructor. Preserve native-order ID/RNG work, authored overrides, and the existing Techno Init_Managers 0x006F3F40 store/child/Reveal transaction.

Move constructor tests into their private module, retain all 15 existing behavioral tests, add a component-route matrix and retain the raw-constructor ownership guard without a brittle call-count assertion. Repair retail factory setup to publish completion through its real owner and verify held identity across rejected and accepted placement.

Validation: constructor16passed; full8482passed0failed80ignored; retail Dustbowl and Hills passed; Clippy and production check succeeded. Rust regression evidence only; existing authored Ship/ammo/Fly differences are preserved without claiming native parity.
